refactor: WriringList 파일 구조 재정리(#47) - WritingLIst 폴더로 WritingList.jsx, WritingList.css 옮김 - WritingItem 컴포넌트를 list 폴더에서 components 폴더로 옮김
